Add a key event queue
Reading input stream can sometimes return a sequence of several key
events. This can happen if we have not been reading the input stream for
some time, for example if the process was stuck because of a long
preview function.

Two modifications are required to handle these key sequences:
- First we have to split this sequence in atomic keys, taking care of
  events that might be represented by more than one byte.
- And next we have to store extra keys for later, so we don't lose any
  input event.

Fixes #99

Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>